1_fargate_task/requirements.txt
newspaper3k
warcio
lxml_html_clean
boto3